- Test for fetching user by ID - Adding new 404 logic - More test coverage for unhappy path

+++ UserManagement.Services/Implementations/UserService.cs
@@ -13,7 +13,7 @@ public class UserService(IDataContext dataAccess) : IUserService
         return dataAccess.GetAll<User>().Where(x => x.IsActive == isActive);
     }
 
-    public User? GetUserById(int id)
+    public User? GetUserById(long id)
     {
         return dataAccess.GetAll<User>().Where(x => x.Id == id).FirstOrDefault();
     }
